(MODULES-10953) Update metadata.json and pdk version
To avoid having to update this everytime we release a new agent
platform, it should be enough to specify the supported OS, without
specific versions. It is assumed that for each OS in metadata.json, the
versions supported are the same as what the agent itself supports.

Note: Puppet does not 'oficially' support ZFS on Linux (meaning we do
not test against Linux configurations and we don't commit to solve and
issues related to ZoL), so remove Linux platforms from metadata.json.
However, we happily accept community contributions that do not break any
Solaris functionality.